requirements.txt
click
pyyaml
rich 